Fix default cursor color
* Set default cursor color by inheriting from current frame's cursor color.
* Respect default cursor color specified in `evil-refresh-cursor'.

This is an attempt to address the problem introduced in "06ed09b". The problem
is that if `evil-emacs-state-cursor' is set, the cursor color changes when
switching to Evil Emacs state which is good; but, the cursor color will be
locked to whatever color specified in `evil-emacs-state-cursor' and will not
change back to the original color after exited from Emacs state.

The problem is two sided, first, since there was no default cursor color, there
was no way for the cursor color to "change back". Second, even if we set the
default cursor color, the function `evil-refresh-cursor' didn't respect the
default color, i.e., it still gets the cursor color from current frame which has
been changed to the color specified in `evil-emacs-state-cursor'.